@@ -2,7 +2,7 @@ module DefaultPageContent
   extend ActiveSupport::Concern
 
   included do
-    before_filter :set_page_defaults
+    before_action :set_page_defaults
   end
 
   def set_page_defaults
