@@ -2,7 +2,7 @@ module DeviseWhitelist
   extend ActiveSupport::Concern
 
   included do
-    before_filter :configure_permitted_parameters, if: :devise_controller?
+    before_action :configure_permitted_parameters, if: :devise_controller?
   end
 
   def configure_permitted_parameters
