@@ -2,7 +2,7 @@ module SetSource
   extend ActiveSupport::Concern
 
   included do
-    before_filter :set_source
+    before_action :set_source
   end
 
   def set_source
